[Project] Modified start script
To make development faster, I have started using the script wihtin
start_scripts/ in order to simply restart the service when I make
changes instead of having to find and kill the process each time.

This commit simply adds my network interface to the script and makes it
run the debug daemon in the upper directory.

It might be worthwhile doing something to this file to prevent git from
picking up any changes as each developer will have a different interface
variable.