[Feature] Added libconfig functionality to project
It is now possible to read in a config file when starting the onvif_svrd
daemon instead of passing in the configuration as command line
arguments.

The basic functionality works the same as before, where if there is a
specific configuration missing from the file, it will load in a default
value where available or quit the daemon.