[FEATURE] Made MQTT functionality persistent
Before this change, the MQTT client was opened and closed on a
message-by-message basis. Now the connection is kept open in each
instance of the ServiceContext class.

I have noticed that if no messages are sent for a while the MQTT handler
starts to spit out errors, so more investigation is needed.