[FEATURE] Added multiple stream compatability
It is now possible to add multiple video streams to a single instance
of the onvif daemon, this has been done to allow support for both a
left and right monitor stream to be sent.

From now on it is advised to simply run the daemon from the
command line without any parameters as the config file will
manage all parameters required by the daemon.

NOTE: You will be required to change the interfaces within the
config file as they will change depending on the device it is running
on.

[ci skip]