[FEATURE] Initial commit of deb build
Some changes to the source and a new script has been added to the
project which will allow users to build a .deb file for easy
installation. Newer changes to the project will add the wsdd daemon to
the same package installation.

In future some additional config will have to be integrated into the
installation script, for example auto selection of the active network
interface