[FEATURE] Added RTSP Configuration
It is now possible to configure the RTSP stream in the config file.
users are able to specify the pipeline, network ports to listen on and
enable/disable the test stream.

This new functionality has been written so more than two RTSP streams
can be configured.

Future changes to the code will allow users to use the test streams by
default, or connect to a running UDP stream and re-package as a RTSP
stream.  Some refactoring to the config file will also be needed later
on.