[FEATURE] Migrated strem threads to ThreadWarden
All detected and valid streams will spin up their own instance of
ThreadWarden.  When the progam is closed they will be correctly handled.

To get this to work with no errors I am using a C style pointer for the
main GStreamer loop.. this is because using the wrapper was causing
issues at runtime.